feat: tag microsoft.com links in frontmatter, config and embeds with MVP id
The rehype-mvp-url plugin only tags Microsoft links inside rendered
markdown bodies, so links that live in frontmatter, static config or
Astro component props were missed.

Extract the tagging logic into a shared lib/mvp-url.ts (tagMvpUrl) and
apply it to:
- speaking `url`, blog `sourceUrl`, note `externalUrl`, acknowledgement
  `url` (via zod transforms, so every consumer gets the tagged URL)
- the Microsoft MVP badge in config/personal.ts
- the SocialEmbed component's linkUrl

Acknowledgement `licenceUrl` (a generic licence-terms page) is left
untagged on purpose. tagMvpUrl is a no-op for non-Microsoft URLs, so the
transforms are safe on fields that usually point elsewhere.

Verified against the built output: previously-untagged Microsoft links
now carry WT.mc_id; licenceUrl stays untouched; build and lint pass.
